Synchro with last fpc 3.2.2. and last fpGUI commits.

+++ src/fpGUI-develop_dynX11/fpg_x11.pas
@@ -1106,8 +1106,8 @@ var
   s: TfpgString;
   itm: TfpgMimeDataItem;
 begin
-  FDNDVersion := min(FPG_XDND_VERSION, (ev.xclient.data.l[1] and $FF000000) shr 24);
-
+   FDNDVersion := min(Int64(FPG_XDND_VERSION), Int64((ev.xclient.data.l[1] and $FF000000) shr 24));
+ 
   {$IFDEF DNDDEBUG}
   DebugLnFmt('  ver(%d) check-XdndTypeList(%s) data=%xh,%d,%d,%d,%d',
       [ FDNDVersion,
